requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
Click==7.0
cmake==3.18.4.post1
Django==3.1.4
djangorestframework==3.12.2
dlib==19.19.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==2.10
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
# pkg-resources==0.0.0
pytz==2020.4
requests==2.25.0
sqlparse==0.4.1
tqdm==4.43.0
urllib3==1.26.2
gunicorn==20.0.4
